Changed the default number of significant digits for BigNumbers from 20 to 64.

+++ test/function/construction/bignumber.test.js
@@ -63,7 +63,7 @@ describe('bignumber', function() {
 
     // restore default precision
     math.config({
-      precision: 20
+      precision: 64
     });
   });
